perf: Query limit filter in unitstore.Refs
If one of our filters is a `limiter`, then we can avoid querying additional
unit stores if our limit is reached. On a large synthetic repo (500 units) this
took query time from 15s to 6s. This was an FS backed store through sourcegraph
on my MBA. The limit used is 2. Most of the 6s time is dominated by
unmarshalling the index, rather than the queries.