store: In memory cache for IndexedTreeStore
IndexedTreeStore contains a few indexes used by most queries. In some repos
these can be large and the time to read, gunzip and unmarshal can be costly.
This commit adds a simple in-memory cache around the indexes. The places where
we `get` and `put` from the cache are only done where the index is read only, so
we do not have to concern ourselves with making the Indexes handle concurrent
writes.

The implementation is currently lacking cache eviction and invalidation. A
future commit will implement that.

In a synthetic repo on my MBA some queries took 6s, most of that time was
dominated by loading indexes in `IndexedTreeStore`. In particular I generated a
repo with

  $ srclib gen-data urefs -v -u 1000 -f 10 -r 'tmp/genrepo11' --ndefs 10 --gen-source

then did the `ListExamples` operation in sourcegraph.