Remove the "srclib toolchain install-std" command entirely.
Keeping a notion of "standard" or "official" toolchains is bad because it makes people
think that srclib is less extensible than it really is.

Also, some toolchains are better than others (e.g. srclib-go and srclib-java) while
others are not of the same quiality yet, so calling them standard is a bit strange.