store.defQueryTreeIndex: Prevent possible nil pointer dereference panic.
There are situations where x.mt.t is nil, so doing node, i := x.mt.t.IndexedTraverse([]rune(q)) results in a nil pointer dereference panic.

This is a quick fix to prevent a panic. A proper fix will be to avoid calling getByQuery when the mafsaTable is not built/read.