cli: revert a part of PR #212
This change reverts a part of PR #212 / 3dcee737b542f81af862a93449eb9deb855cce3f which
removed a hack on the pretense that it worked fine locally without it. Although it is
true that it works fine without it locally (e.g. running `srclib config -m docker` works),
we are seeing failures to build the github.com/golang/go repository due to file permission
errors:

  ##### Building Go bootstrap tool. cmd/dist
  # _/src/src/cmd/dist cannot create cmd/dist/dist: Permission denied
  PreConfigCommands: command "\nif [ -d /home/srclib ]; then\ncurl -L 'https://storage.googleapis.com/golang/go1.4.1.linux-amd64.tar.gz' > /tmp/go1.4.tgz &&\nmkdir /home/srclib/go1.4 &&\ntar -xf /tmp/go1.4.tgz -C /home/srclib/go1.4 --strip-components=1\nfi &&\ncd src\n./make.bash\n": exit status 2

Suspected is that older Docker versions handle the file permissions differently, thus it works
locally but not in prod.