allow importing when no toolchains are present
Removes the SRCLIB_NO_TOOLCHAINS env var, which permitted `srclib
store import` operations (and other things that depended on already
produced data) even when no toolchains were present. This behavior is
now the standard behavior.

This could make it slightly more confusing if you are trying to debug
why a scanned source unit is not being passed to a grapher (e.g., if
you specify the wrong ToolRef to use to graph the source unit), but it
comes at a significant decrease in complexity overall.