remove editor plugins (support code, documentation)
The srclib editor plugins have no longer been supported for a while.
This commit removes the srclib internal CLI commands that editor
plugins relied on, as well as the documentation describing editor
plugin-related information. Because the toolchains and data structures
are not as useful to a general audience now that the editor plugins
are removed, those documentation pages are also removed.

Releasing much-improved editor plugins is planned for the future; stay
tuned!