add GraphMultiUnitsRule to graph multiple units of same type together
This allows language toolchains to specify a 'graph-all' op for a source
unit, which will combine all units of the same type into one rule and
call the language toolchain to graph them together.

This feature can be leveraged to avoid duplicate work done for loading
common dependencies of the source units. This can lead to significant
reduction in graphing time.